Update: Handles Errors and Changes

+++ Todo-Backend/src/service/TodoService.ts
@@ -1,3 +1,4 @@
+import { NotFoundError } from "../errors/NotFoundError";
 import * as todoRepo from "../Repositories/TodoRepo";
 import { Todo } from "../model/TodoModel";
 
@@ -6,21 +7,29 @@ export const getAllTodos = () => {
 };
 
 export const getUserTodos = (id: number) => {
-  return todoRepo.getUserTodos(id);
-}
+  const todos = todoRepo.getUserTodos(id);
+  if (!todos) throw new NotFoundError("Todos not found");
+  return todos;
+};
 
 export const addTodo = (todo: Todo) => {
   return todoRepo.addTodo(todo);
 };
 
-export const getTodoById = (id: number,userid : number) => {
-  return todoRepo.getTodoById(id,userid);
+export const getTodoById = (id: number, userid: number) => {
+  const todos = todoRepo.getTodoById(id, userid);
+  if (!todos) throw new NotFoundError("Todo not found");
+  return todos;
 };
 
-export const updateTodo = (id:number,todo: Todo) => {
-  return todoRepo.updateTodo(id,todo);
+export const updateTodo = (id: number, todo: Todo) => {
+  const foundTodo = todoRepo.getTodoById(id, todo.userid);
+  if (!foundTodo) throw new NotFoundError("Todo not found");
+  return todoRepo.updateTodo(id, todo);
 };
 
-export const deleteTodo = (id: number) => {
+export const deleteTodo = (id: number, userid: number) => {
+  const todo = todoRepo.getTodoById(id, userid);
+  if (!todo) throw new NotFoundError("Todo not found");
   return todoRepo.deleteTodo(id);
 };
